+++ Chapter05/pom.xml
@@ -2,9 +2,13 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.zdong</groupId>
+  <parent>
+    <artifactId>Chapter</artifactId>
+    <groupId>com.course.testNg</groupId>
+    <version>1.0</version>
+  </parent>
+
   <artifactId>Chapter05</artifactId>
-  <version>1.0</version>
   <packaging>jar</packaging>
 
   <name>Chapter05</name>
@@ -18,13 +22,13 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>3.8.2</version>
+      <version>4.13.2</version>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.testng</groupId>
       <artifactId>testng</artifactId>
-      <version>6.8.7</version>
+      <version>7.1.0</version>
     </dependency>
   </dependencies>
 </project>
