+++ Chapter05/src/main/java/com/course/testNG/BasicAnnotations.java
@@ -1,0 +1,16 @@
+package com.course.testNG;
+
+import org.testng.annotations.Test;
+
+public class BasicAnnotations {
+    @Test
+    public void testCase01(){
+        System.out.println("这是测试case01");
+    }
+    @Test
+    public void testCase02(){
+        System.out.println("这是测试case02");
+    }
+
+
+}
