+++ Chapter05/src/main/java/com/course/testNG/multThread/TestMultThreadXml.java
@@ -1,0 +1,21 @@
+package com.course.testNG.multThread;
+
+import org.testng.annotations.Test;
+
+public class TestMultThreadXml {
+
+    @Test
+    public void testo1(){
+        System.out.printf("Thread Id is %d%n",Thread.currentThread().getId());
+    }
+
+    @Test
+    public void test02(){
+        System.out.printf("Thread Id is %d%n",Thread.currentThread().getId());
+    }
+
+    @Test
+    public void test03(){
+        System.out.printf("Thread Id is %d%n",Thread.currentThread().getId());
+    }
+}
