+++ Chapter05/src/main/java/com/course/testNG/suite/LoginTest.java
@@ -1,0 +1,10 @@
+package com.course.testNG.suite;
+
+import org.testng.annotations.Test;
+
+public class LoginTest {
+    @Test
+    public void Login(){
+        System.out.println("登录成功");
+    }
+}
