+++ Chapter05/src/main/java/com/course/testNG/suite/PayTest.java
@@ -1,0 +1,10 @@
+package com.course.testNG.suite;
+
+import org.testng.annotations.Test;
+
+public class PayTest {
+    @Test
+    public void Pay(){
+        System.out.println("支付宝到账100万");
+    }
+}
