--- Chapter05/src/main/java/com/zdong/BasicAnnotations.java
@@ -1,16 +1,0 @@
-package com.zdong;
-
-import org.testng.annotations.Test;
-
-public class BasicAnnotations {
-    @Test
-    public void testCase01(){
-        System.out.println("这是测试case01");
-    }
-    @Test
-    public void testCase02(){
-        System.out.println("这是测试case02");
-    }
-
-
-}
