--- Chapter05/src/main/java/com/zdong/TestExceptions.java
@@ -1,18 +1,0 @@
-package com.zdong;
-
-import org.testng.annotations.Test;
-
-public class TestExceptions {
-    @Test(expectedExceptions = RuntimeException.class)
-    public void RuntimeExceptionFailed(){
-        System.out.println("这是一个失败的异常测试");
-    }
-
-    @Test(expectedExceptions = RuntimeException.class)
-    public void RuntimeExceptionSucceed(){
-        System.out.println("这是一个成功的异常测试");
-        throw new RuntimeException();
-    }
-
-
-}
