--- Chapter05/src/main/java/com/zdong/groups/GroupOnClass1.java
@@ -1,13 +1,0 @@
-package com.zdong.groups;
-
-import org.testng.annotations.Test;
-
-@Test(groups = "stu")
-public class GroupOnClass1 {
-    public void testStu1(){
-        System.out.println("class 01 的 11");
-    }
-    public void testStu2(){
-        System.out.println("class 01 的 22");
-    }
-}
