--- Chapter05/src/main/java/com/zdong/groups/GroupOnclass2.java
@@ -1,14 +1,0 @@
-package com.zdong.groups;
-
-import org.testng.annotations.Test;
-
-@Test(groups = "stu")
-public class GroupOnclass2 {
-
-    public void testStu1(){
-        System.out.println("class 02 的 11");
-    }
-    public void testStu2(){
-        System.out.println("class 02 的 22");
-    }
-}
