--- Chapter05/src/main/java/com/zdong/suite/LoginTest.java
@@ -1,10 +1,0 @@
-package com.zdong.suite;
-
-import org.testng.annotations.Test;
-
-public class LoginTest {
-    @Test
-    public void Login(){
-        System.out.println("登录成功");
-    }
-}
