--- Chapter05/src/main/java/com/zdong/suite/PayTest.java
@@ -1,10 +1,0 @@
-package com.zdong.suite;
-
-import org.testng.annotations.Test;
-
-public class PayTest {
-    @Test
-    public void Pay(){
-        System.out.println("支付宝到账100万");
-    }
-}
