--- Chapter05/src/main/java/com/zdong/suite/SuiteConfig.java
@@ -1,16 +1,0 @@
-package com.zdong.suite;
-
-import org.testng.annotations.AfterSuite;
-import org.testng.annotations.BeforeSuite;
-
-public class SuiteConfig {
-    @BeforeSuite
-    public void BeforeSuite(){
-        System.out.println("这是beforesuit");
-    }
-
-    @AfterSuite
-    public void AfterSuite(){
-        System.out.println("这是aftersuite");
-    }
-}
