+++ Chapter05/src/main/resources/paramter.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<suite name="test">
+    <test name="paramter">
+        <classes>
+            <parameter name="name" value="Jenny"/>
+            <parameter name="age" value="18"/>
+
+            <class name = "com.course.testNG.paramter.TestParam"/>
+        </classes>
+    </test>
+
+</suite>
