+++ Chapter06/src/main/java/com/course/testNg/App.java
@@ -1,0 +1,13 @@
+package com.course.testNg;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
