--- test_con.txt
@@ -1,2 +1,0 @@
-this is the dev
-这是需要合并的内容
